ath79: add support for Mikrotik RouterBOARD 2011
This commit restores support for all RB2011 versions which were
supported in the removed ar71xx target. It also cleanly separates
all available RB2011 versions and it should fix uncertain status
of USB and SFP support.

Ethernet 6-10 LEDs are not working because AR8327 driver does not
support exposing LEDs when using device tree (relevant code is not
present in the device tree parser).